# perf(fastrand): optimize Read 