--- a/.github/workflows/feishu-notify.yml
+++ b/.github/workflows/feishu-notify.yml
@@ -26,8 +26,8 @@ jobs:
             See: https://github.com/${{ github.repository }}/pull/${{ github.event.pull_request.number }}
 
         run: |
-          sudo apt update -y
-          sudo apt install wget
+          sudo apt update
+          sudo apt install wget -y
           wget -q https://github.com/xiachufang/actions-feishu/releases/download/${{ env.ACTIONS_FEISHU_TAG }}/linux-amd64-actions-feishu.tar.gz
           tar zxf linux-amd64-actions-feishu.tar.gz feishu
           ./feishu
@@ -52,7 +52,7 @@ jobs:
 
         run: |
           sudo apt update -y
-          sudo apt install wget
+          sudo apt install wget -y
           wget -q https://github.com/xiachufang/actions-feishu/releases/download/${{ env.ACTIONS_FEISHU_TAG }}/linux-amd64-actions-feishu.tar.gz
           tar zxf linux-amd64-actions-feishu.tar.gz feishu
           ./feishu