--- a/.github/workflows/feishu-notify.yml
+++ b/.github/workflows/feishu-notify.yml
@@ -51,7 +51,7 @@ jobs:
             See: https://github.com/${{ github.repository }}/issues/${{ github.event.issue.number }}
 
         run: |
-          sudo apt update -y
+          sudo apt update
           sudo apt install wget -y
           wget -q https://github.com/xiachufang/actions-feishu/releases/download/${{ env.ACTIONS_FEISHU_TAG }}/linux-amd64-actions-feishu.tar.gz
           tar zxf linux-amd64-actions-feishu.tar.gz feishu